All data saved!

+++ backend/services/auth.service.js
@@ -92,9 +92,11 @@ class AuthService {
     if (!existUser) {
       throw BaseError.BadRequest("User not found!");
     }
-    existUser.image.length > 0 && image
-      ? FileService.delete(existUser.image)
-      : null;
+    if (image && existUser.image && existUser.image.length > 0) {
+      existUser.image.length > 0 && image
+        ? FileService.delete(existUser.image)
+        : null;
+    }
     const fileName = image ? FileService.save(image) : existUser.image;
     const userPayload = tokenService.validateAccessToken(accessToken);
